what

+++ main.py
@@ -10,11 +10,11 @@ class CameraImage:
     self.file_name = file_name
 
   def averageLightValue(self):
-    return cv2.mean(self.img_gray)
+    return int(cv2.mean(self.img_gray))
 
   def blurrinessValue(self):
     img_lap = cv2.Laplacian(self.img_gray, cv.CV_16S, ksize=1, scale=1, delta=0)
-    return cv2.reduce(cv2.reduce(img_lap, 0, cv.CV_REDUCE_MAX), 1, cv.CV_REDUCE_MAX)[0][0]
+    return int(cv2.reduce(cv2.reduce(img_lap, 0, cv.CV_REDUCE_MAX), 1, cv.CV_REDUCE_MAX)[0][0])
     
 
 class Camera:
